Bug fix on SYNTHESIS printf and fine tune timing
Added SYNTHESIS macro, 1 for building on FPGA, 0 for
building simulation.
Fine tune the timing on task sleeping, for fpga build, since we depends
on the fpga clocks, we need to slow down the switching time, otherwise
it LEDS will blinks too fast. For debuging build, I just make sure
the task a switched as fast as possible to demontrate multitasking
on AFTx07

Sign-off-by: Tin Tran <trantin0815@gmail.com>